# Optimize allocation behavior of the autoscaler