<!-- Please include the 'why' behind your changes if no issue exists -->
## Proposed Changes

This optimizes the autoscaler's allocation behavior by
1. Collapsing metric record calls, since that path is kind of expensive
2. Avoiding to build a new logger and instead only adding the metric key where relevant
3. Avoiding to build new loggers for the stable and panic debug logs
4. Avoiding to call formatting logs by checking the log levels first

With the benchmark shipped in #10374, this results in the following gains:

```
benchmark                  old ns/op     new ns/op     delta
BenchmarkAutoscaler-16     12246         3323          -72.86%

benchmark                  old allocs     new allocs     delta
BenchmarkAutoscaler-16     83             22             -73.49%

benchmark                  old bytes     new bytes     delta
BenchmarkAutoscaler-16     6131          920           -84.99%
```

Pulling in https://github.com/knative/pkg/pull/1964 will further reduce this to:

```
benchmark                  old ns/op     new ns/op     delta
BenchmarkAutoscaler-16     12246         2069          -83.10%

benchmark                  old allocs     new allocs     delta
BenchmarkAutoscaler-16     83             10             -87.95%

benchmark                  old bytes     new bytes     delta
BenchmarkAutoscaler-16     6131          632           -89.69%
```

/assign @vagababov