@vagababov 2020-12-15
/test pull-knative-serving-istio-stable-no-mesh